% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/aggregateRunoff.r
\name{aggregateRunoff}
\alias{AggregateRunoff}
\alias{aggregateRunoff}
\title{Aggregate Runoff}
\usage{
AggregateRunoff(ncFile, catchmentPolygons, runoffVar, useWeights = F,
  sumData = T, startDate = NULL, leapDays = F, by = "day",
  fname = NULL, convertToDischarge = T)
}
\arguments{
\item{ncFile}{NetCDF runoff file to be processed}

\item{catchmentPolygons}{catchments to be used for aggregation}

\item{runoffVar}{netCDF variable that the runoff data is stored in}

\item{startDate}{simulation start date, needed for aggregating by month and rownames. Default is NULL and ok if only using daily data}

\item{by}{time step of runoff. Can be "day", or "month". Default is day. Used for unit conversion and rownames}

\item{fname}{filename to write data to. Default is NULL. If not NULL, data will be written to text file in format fname.txt}

\item{leapDay}{set to False if leapdays are not included in simulation. Used for setting rownames of daily runoff. Default is F}
}
\value{
Dataframe of runoff aggregated by catchments and time
}
\description{
Reads in runnoff NetCDFs and generates matrix of runoff aggregated by catchment and time stepusing the extract() function
}
\examples{
surfaceRunoff <- AggregateRunoff(ncFile=paste(ncdir, "/",  surfaceNcName, sep=""), catchmentPolygons=catchmentsToUse, runoffVar=surfaceVarName, startDate=simStartDate, by=timeStep)
}

