% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/aggregateSnow.r
\name{AggregateSnow}
\alias{AggregateSnow}
\title{Aggregate Runoff}
\usage{
AggregateSnow(defaults = setupList, catchmentPolygons, snowFileName = NULL,
  msroFileName = NULL, ncDir = NULL, runoffVar = NULL,
  useWeights = NULL, startDate = NULL, leapDays = F, by = NULL,
  fname = NULL)
}
\arguments{
\item{catchmentPolygons}{catchments to be used for aggregation}

\item{runoffVar}{netCDF variable that the runoff data is stored in}

\item{startDate}{simulation start date, needed for aggregating by month and rownames. Default is NULL and ok if only using daily data}

\item{by}{time step of runoff. Can be "day", or "month". Default is day. Used for unit conversion and rownames}

\item{fname}{filename to write data to. Default is NULL. If not NULL, data will be written to text file in format fname.txt}

\item{ncFile}{NetCDF runoff file to be processed}

\item{leapDay}{set to False if leapdays are not included in simulation. Used for setting rownames of daily runoff. Default is F}
}
\value{
Dataframe of runoff aggregated by catchments and time
}
\description{
Reads in runnoff NetCDFs and generates matrix of runoff aggregated by catchment and time stepusing the extract() function
}
\examples{
surfaceRunoff <- AggregateRunoff(ncFile=paste(ncDir, "/",  surfaceNcName, sep=""), catchmentPolygons=catchmentsToUse, runoffVar=surfaceVarName, startDate=simStartDate, by=timeStep)
}

