% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/gaugeDataFunctions.r, R/subsetShapesFunction.r
\name{cleanDat}
\alias{GetShapesById}
\alias{GetShapesInBounds}
\alias{cleanDat}
\title{Fill Data Gaps}
\usage{
cleanDat(dat, simStartDate, simEndDate)

GetShapesInBounds(shapeFrame, hucCodes, by = "HUC")

GetShapesById(shapeFrame, IDs)
}
\arguments{
\item{dat}{dataframe of guage data to be cleaned and filled}

\item{simStartDate}{date of the beginning of the simulation}

\item{simEndDate}{date of the end of the simulation}

\item{dat}{dataframe of guage data to be cleaned and filled}

\item{simStartDate}{date of the beginning of the simulation}

\item{simEndDate}{date of the end of the simulation}

\item{dat}{dataframe of guage data to be cleaned and filled}

\item{simStartDate}{date of the beginning of the simulation}

\item{simEndDate}{date of the end of the simulation}
}
\value{
Dataframe of stream gauge data with missing values filled with NAs

Dataframe of stream gauge data with missing values filled with NAs

Dataframe of stream gauge data with missing values filled with NAs
}
\description{
Fills gaps in daily stream gaugedata with NA so that statistics can later be calculated with data. Function only used by getGaugeData

Fills gaps in daily stream gaugedata with NA so that statistics can later be calculated with data. Function only used by getGaugeData

Fills gaps in daily stream gaugedata with NA so that statistics can later be calculated with data. Function only used by getGaugeData
}
\examples{
cleanDat(dat, simStartDate, simEndDate)
cleanDat(dat, simStartDate, simEndDate)
cleanDat(dat, simStartDate, simEndDate)
}

