% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/gaugeDataFunctions.r
\name{GetGaugeData}
\alias{GetGaugeData}
\title{USGS Gauge Data Fetcher}
\usage{
GetGaugeData(edges, gauges, aggregateByMonth = T, maxDist = 0.005,
  idField = edgeIdField, checkGauges = F)
}
\arguments{
\item{edges}{SpatialLinesDataFrame of edges. Used to find gauges that correspond to stream network edges}

\item{gauges}{SpatialPointsDataFrame of NWIS stream gauges. Download shapefile from nwis website}

\item{aggregateByMonth}{logical of whether to aggregate stream gauge data by month, used for comparing to monthly simulation data. Defaults to T}

\item{maxDist}{maximum distance that a stream gauge point can be from a line and still snap to it. Can be calibrated for better snapping accuracy. Units in decimal degrees (I think). Defaulted to .005 decimal degrees}

\item{idField}{field of edge data frame that contains the identifying field to be used for relating stream gauge points to edges. Defaults to edgeIdField, a parameter set at the beginning of routeWater.r script}

\item{checkGauges}{logical indicating whether to plot and manually approve of gauge snap befor processing. Defaulted to FALSE}
}
\value{
List of dataframes for each stream gauge
}
\description{
Finds USGS stream gauges close to edges, downloads, cleans, and returns a list of dataframes for each gauge
}
\examples{
gaugeData <- GetGaugeData(edgesInBounds, nwisGauges, aggregateByMonth=aggregateGaugeDataByMonth, checkGauges=T)
}

