% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/subsetShapesFunction.r
\name{getOrder}
\alias{findAllParents}
\alias{getOrder}
\title{Get Order}
\usage{
getOrder(shapeFrame, id, idField = edgeIdField, orderField = edgeOrderField)

findAllParents(shapeFrame, ID, createArray = T, idField = edgeIdField,
  nextDownField = edgeNextDownField, orderField = edgeOrderField)
}
\arguments{
\item{shapeFrame}{a spatialFrame of edges or catchments}

\item{id}{ID of the feature that you are trying to find the order of}

\item{idField}{the field that shape IDs are stored in. Defaulted to edgeIdField, a parameter set in the beginning of routeWater.r}

\item{orderField}{the field that stream order of the shapes are stored in. This is a field created automatically in ArcGIS processing. Defaulted to edgeNextDownField, a parameter set in the beginning of routeWater.}

\item{shapeFrame}{a spatialFrame of edges or catchments}

\item{id}{ID of the feature that you are trying to return all the parents of}

\item{idField}{the field that shape IDs are stored in. Defaulted to edgeIdField, a parameter set in the beginning of routeWater.r}

\item{orderField}{the field that stream order of the shapes are stored in. This is a field created automatically in ArcGIS processing. Defaulted to edgeNextDownField, a parameter set in the beginning of routeWater.}
}
\value{
the order of the edge being used

the order of the edge being used
}
\description{
Returns the order of an edge or catchment given it's id

Recursive fucntion that returns all the parents of an edge or catchment given it's ID
}
\examples{
getOrder(edgesInBounds, "34342")
getOrder(edgesInBounds, "34342")
}

