% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/subsetShapesFunction.r
\name{getParents}
\alias{getParents}
\title{Get Parents}
\usage{
getParents(shapeFrame, id, idField = edgeIdField,
  nextDownField = edgeNextDownField)
}
\arguments{
\item{shapeFrame}{a spatialFrame of edges or catchments}

\item{id}{ID of the feature that you are trying to find the parents of}

\item{idField}{the field that shape IDs are stored in. Defaulted to edgeIdField, a parameter set in the beginning of routeWater.r}

\item{nextDownField}{the field that nextDownId of the shapes are stored in. This is a field created automatically in ArcGIS processing. Defaulted to edgeNextDownField, a parameter set in the beginning of routeWater.r}
}
\value{
the parents of the edge being used
}
\description{
Returns the parents of an edge or catchment given it's id
}
\examples{
getParents(edgesInBounds, "34342")
}

