% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/notifyMe.r
\name{notifyMe}
\alias{notifyMe}
\title{Email Notifications}
\usage{
notifyMe(subject = "Job Finished", body = " ",
  email = "jeradhoy@gmail.com")
}
\arguments{
\item{subject}{character string for subject of email. Defaults to "Job Finished"}

\item{body}{optional character string for body of email. Defaults to nothing}

\item{email}{character string for email address to send to/from. Defaults to "jeradhoyATgmail.com"}
}
\value{
None
}
\description{
Emails a message from R when called, useful for code with long computation time
}
\examples{
notifyMe()
notifyMe("Routing Finished")
notifyMe("Routing Finished", "Finished routing historical dayment data", emailaddressATexample.com)
notifyMe(paste("Routing Finished at", Sys.time()))
}

