% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routeWaterCpp.r, R/routingFunctions.r
\name{RouteWater}
\alias{RouteWater}
\alias{RouteWaterCpp}
\title{Routes surface and subsurface water through river network}
\usage{
RouteWaterCpp(edges, catchments, Rsurf, Rsub, defaults = setupList,
  spinUpCycles = 0, spinUpYears = 10, debugMode = F, by = "day",
  widthCoeffs = c(0.3, 0.6), manningN = 0.07, slopeMin = 0.01,
  aCoeffCoeff = 3, outputExtraVars = T, etaInt = 10, beaverCoeff = NULL)

RouteWater(edges, catchments, Rsurf, Rsub, spinUpCycles = 0,
  spinUpYears = 10, debugMode = F, by = "day", widthCoeffs = c(0.3,
  0.6), manningN = 0.07, slopeMin = 0.01, aCoeffCoeff = 3,
  outputExtraVars = T, etaInt = 10)
}
\arguments{
\item{dat}{dataframe of guage data to be cleaned and filled}

\item{simStartDate}{date of the beginning of the simulation}

\item{simEndDate}{date of the end of the simulation}

\item{dat}{dataframe of guage data to be cleaned and filled}

\item{simStartDate}{date of the beginning of the simulation}

\item{simEndDate}{date of the end of the simulation}
}
\value{
Dataframe of stream gauge data with missing values filled with NAs

Dataframe of stream gauge data with missing values filled with NAs
}
\description{
Routes surface and subsurface water through river network

Routes surface and subsurface water through river network
}
\examples{
cleanDat(dat, simStartDate, simEndDate)

cleanDat(dat, simStartDate, simEndDate)

}

