% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamTemp2.r, R/streamTemp.r
\name{StreamTemp}
\alias{StreamTemp}
\alias{StreamTempCpp}
\title{Stream Temperature}
\usage{
StreamTempCpp(edges, catchments, RsurfSnow, RsurfNoSnow, Tair, simFlow,
  defaults = setupList, by = "month", outputExtraVars = T,
  debugMode = T, runStart = 1, runStop = NULL, K = 0.1, etaInt = 10,
  outFile = NULL, prof = NULL)

StreamTemp(edges, catchments, RsurfSnow, RsurfNoSnow, Tair, simFlow,
  defaults = setupList, by = "month", outputExtraVars = T,
  debugMode = T, runStart = 1, runStop = NULL, K = 0.1, etaInt = 10,
  outFile = NULL, prof = NULL)
}
\arguments{
\item{outputExtraVars}{}

\item{outputExtraVars}{}
}
\value{
List of matrices of routing variable timeseries.

List of matrices of routing variable timeseries.
}
\description{
Routes runoff through a stream network using various physical relationships.

Routes runoff through a stream network using various physical relationships.
}
\examples{
flow <- RouteWater(edges=edgesInBounds, catchments=catchmentsInBounds, Rsurf=surfaceRunoff,  Rsub=subsurfRunoff, spinUpCycles=gwSpinUpCycles, spinUpYears=spinUpYears, debugMode=F, by=timeStep, widthCoeffs=streamWidthCoeffs, manningN=manningN, slopeMin=slopeMin, aCoeffCoeff=aCoeffCoeff)

flow <- RouteWater(edges=edgesInBounds, catchments=catchmentsInBounds, Rsurf=surfaceRunoff,  Rsub=subsurfRunoff, spinUpCycles=gwSpinUpCycles, spinUpYears=spinUpYears, debugMode=F, by=timeStep, widthCoeffs=streamWidthCoeffs, manningN=manningN, slopeMin=slopeMin, aCoeffCoeff=aCoeffCoeff)

}

