% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/routingFunctions.r
\name{RouteWater}
\alias{RouteWater}
\title{Route Water}
\usage{
RouteWater(edges, catchments, Rsurf, Rsub, spinUpCycles = 0,
  spinUpYears = 10, debugMode = F, by = "day", widthCoeffs = c(0.3,
  0.6), manningN = 0.07, slopeMin = 0.01, aCoeffCoeff = 3,
  outputExtraVars = T)
}
\arguments{
\item{edges}{}

\item{catchments}{}

\item{Rsurf}{}

\item{Rsub}{}

\item{spinUpCycles}{}

\item{spinUpYears}{}

\item{debugMode}{}

\item{by}{}

\item{widthCoeffs}{}

\item{manningN}{}

\item{slopeMin}{}

\item{aCoeffCoeff}{}

\item{outputExtraVars}{}
}
\value{
List of matrices of routing variable timeseries.
}
\description{
Routes runoff through a stream network using various physical relationships.
}
\examples{
flow <- RouteWater(edges=edgesInBounds, catchments=catchmentsInBounds, Rsurf=surfaceRunoff,  Rsub=subsurfRunoff, spinUpCycles=gwSpinUpCycles, spinUpYears=spinUpYears, debugMode=F, by=timeStep, widthCoeffs=streamWidthCoeffs, manningN=manningN, slopeMin=slopeMin, aCoeffCoeff=aCoeffCoeff)
}

