% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/routingFunctions.r
\name{RouteWater}
\alias{RouteWater}
\title{Routes surface and subsurface water through river network}
\usage{
RouteWater(edges, catchments, Rsurf, Rsub, spinUpCycles = 0,
  spinUpYears = 10, debugMode = F, by = "day", widthCoeffs = c(0.3,
  0.6), manningN = 0.07, slopeMin = 0.01, aCoeffCoeff = 3,
  outputExtraVars = T)
}
\arguments{
\item{dat}{dataframe of guage data to be cleaned and filled}

\item{simStartDate}{date of the beginning of the simulation}

\item{simEndDate}{date of the end of the simulation}
}
\value{
Dataframe of stream gauge data with missing values filled with NAs
}
\description{
Routes surface and subsurface water through river network
}
\examples{
cleanDat(dat, simStartDate, simEndDate)
}

