% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/streamTemp.r
\name{StreamTemp}
\alias{StreamTemp}
\title{Stream Temperature}
\usage{
StreamTemp(edges, catchments, RsurfSnow, RsurfNoSnow, Tair, simFlow,
  defaults = setupList, by = "month", outputExtraVars = T,
  debugMode = T, runLength = NULL, K = 0.1)
}
\arguments{
\item{edges}{}

\item{catchments}{}

\item{by}{}

\item{outputExtraVars}{}

\item{debugMode}{}

\item{Rsurf}{}

\item{Rsub}{}

\item{spinUpCycles}{}

\item{spinUpYears}{}

\item{widthCoeffs}{}

\item{manningN}{}

\item{slopeMin}{}

\item{aCoeffCoeff}{}
}
\value{
List of matrices of routing variable timeseries.
}
\description{
Routes runoff through a stream network using various physical relationships.
}
\examples{
flow <- RouteWater(edges=edgesInBounds, catchments=catchmentsInBounds, Rsurf=surfaceRunoff,  Rsub=subsurfRunoff, spinUpCycles=gwSpinUpCycles, spinUpYears=spinUpYears, debugMode=F, by=timeStep, widthCoeffs=streamWidthCoeffs, manningN=manningN, slopeMin=slopeMin, aCoeffCoeff=aCoeffCoeff)
}

